// src/app/core/acl/interface.ts
export enum ROLES {
    SUPER = 'SUPER',         // 系统管理员
    ADMIN = 'ADMIN',         // 区域管事员
    DEPART = 'DEPART',       // 机构管理员
    OPER = 'OPER',           // 运维人员
    PERSONNEL = 'PERSONNEL',  // 机构工作人员
    BUSINESS = 'BUSINESS'  // 商家
}